+++ R/Authenticate.R
@@ -1,9 +1,9 @@
 #' @title Create a credential object to access social media APIs
 #'
-#' @description \code{Authenticate} creates a \code{credential} object that enables R to make authenticated calls to social media
-#' APIs. A \code{credential} object is a S3 object containing authentication related information such as an access 
-#' token or key, and a class name identifying the social media that grants authentication. \code{Authenticate} is the 
-#' first step of the \code{Authenticate}, \code{\link{Collect}} and \code{\link{Create}} workflow.
+#' @description \code{Authenticate} creates a \code{credential} object that enables R to make authenticated calls to 
+#' social media APIs. A \code{credential} object is a S3 object containing authentication related information such as an
+#' access token or key, and a class name identifying the social media that grants authentication. \code{Authenticate} is
+#' the first step of the \code{Authenticate}, \code{\link{Collect}} and \code{\link{Create}} workflow.
 #'
 #' Refer to \code{\link{Authenticate.twitter}}, \code{\link{Authenticate.youtube}} and 
 #' \code{\link{Authenticate.reddit}} for parameters and usage.